Realsense nodelet crash - nodelet::Loader::Impl
The `nodelet::detail::CallbackQueueManager::removeQueue` removes an entry from the list of queues without checking first that the entry exists in the queue. When that happens, the underlying C++ library throws an object not found exception aborting the process.
This change only adds a test to make we only remove a queue that exists.